+++ reader_core/src/pnp/utils.rs
@@ -5,6 +5,6 @@ pub fn title_id() -> u64 {
     unsafe { bindings::host_get_game_title_id() }
 }
 
-pub fn update_version() -> u64 {
+pub fn update_version() -> u16 {
     unsafe { bindings::get_remaster_version() }
 }
